separate communication from controller
