+++ nvflare/apis/controller_spec.py
@@ -267,6 +267,19 @@ class ControllerSpec(ABC):
         pass
 
     @abstractmethod
+    def control_flow(self, abort_signal: Signal, fl_ctx: FLContext):
+        """This is the control logic for the RUN.
+
+        NOTE: this is running in a separate thread, and its life is the duration of the RUN.
+
+        Args:
+            fl_ctx: the FL context
+            abort_signal: the abort signal. If triggered, this method stops waiting and returns to the caller.
+
+        """
+        pass
+
+    @abstractmethod
     def stop_controller(self, fl_ctx: FLContext):
         """Stops the controller.
