+++ nvflare/apis/responder.py
@@ -15,7 +15,6 @@
 from abc import ABC, abstractmethod
 from typing import Tuple
 
-from nvflare.apis.signal import Signal
 
 from .client import Client
 from .fl_component import FLComponent
@@ -89,19 +88,6 @@ class Responder(FLComponent, ABC):
 
         Args:
             fl_ctx: FL context. It must contain 'job_id' that is to be initialized
-
-        """
-        pass
-
-    @abstractmethod
-    def control_flow(self, abort_signal: Signal, fl_ctx: FLContext):
-        """This is the control logic for the RUN.
-
-        NOTE: this is running in a separate thread, and its life is the duration of the RUN.
-
-        Args:
-            fl_ctx: the FL context
-            abort_signal: the abort signal. If triggered, this method stops waiting and returns to the caller.
 
         """
         pass
