+++ nvflare/app_common/ccwf/client_ctl.py
@@ -192,7 +192,7 @@ class ClientSideController(Executor, TaskController):
                 return
             report = self._get_status_report()
             if not report:
-                self.log_info(fl_ctx, "nothing to report this time")
+                self.log_debug(fl_ctx, "nothing to report this time")
                 return
             self._add_status_report(report, fl_ctx)
             self.last_status_report_time = report.timestamp
