+++ nvflare/private/fed/server/server_runner.py
@@ -20,6 +20,7 @@ from nvflare.apis.event_type import EventType
 from nvflare.apis.fl_component import FLComponent
 from nvflare.apis.fl_constant import FilterKey, FLContextKey, ReservedKey, ReservedTopic, ReturnCode
 from nvflare.apis.fl_context import FLContext
+from nvflare.apis.impl.wf_comm_server import WFCommServer
 from nvflare.apis.server_engine_spec import ServerEngineSpec
 from nvflare.apis.shareable import ReservedHeaderKey, Shareable, make_reply
 from nvflare.apis.signal import Signal
@@ -122,12 +123,15 @@ class ServerRunner(TBI):
             wf = self.config.workflows[self.current_wf_index]
             try:
                 with self.engine.new_context() as fl_ctx:
-                    self.log_info(fl_ctx, "starting workflow {} ({}) ...".format(wf.id, type(wf.responder)))
+                    self.log_info(fl_ctx, "starting workflow {} ({}) ...".format(wf.id, type(wf.controller)))
 
                     fl_ctx.set_prop(FLContextKey.WORKFLOW, wf.id, sticky=True)
-                    wf.responder.initialize_run(fl_ctx)
 
-                    self.log_info(fl_ctx, "Workflow {} ({}) started".format(wf.id, type(wf.responder)))
+                    wf.controller.set_communicator(WFCommServer(), fl_ctx)
+                    wf.controller.communicator.initialize_run(fl_ctx)
+                    wf.controller.start_controller(fl_ctx)
+
+                    self.log_info(fl_ctx, "Workflow {} ({}) started".format(wf.id, type(wf.controller)))
                     self.log_debug(fl_ctx, "firing event EventType.START_WORKFLOW")
                     self.fire_event(EventType.START_WORKFLOW, fl_ctx)
 
@@ -137,7 +141,7 @@ class ServerRunner(TBI):
                         self.current_wf = wf
 
                 with self.engine.new_context() as fl_ctx:
-                    wf.responder.control_flow(self.abort_signal, fl_ctx)
+                    wf.controller.control_flow(self.abort_signal, fl_ctx)
             except Exception as e:
                 with self.engine.new_context() as fl_ctx:
                     self.log_exception(fl_ctx, "Exception in workflow {}: {}".format(wf.id, secure_format_exception(e)))
@@ -155,7 +159,8 @@ class ServerRunner(TBI):
 
                     self.log_info(fl_ctx, f"Workflow: {wf.id} finalizing ...")
                     try:
-                        wf.responder.finalize_run(fl_ctx)
+                        wf.controller.stop_controller(fl_ctx)
+                        wf.controller.communicator.finalize_run(fl_ctx)
                     except Exception as e:
                         self.log_exception(
                             fl_ctx, "Error finalizing workflow {}: {}".format(wf.id, secure_format_exception(e))
@@ -304,7 +309,7 @@ class ServerRunner(TBI):
                 )
                 with self.wf_lock:
                     if self.current_wf:
-                        self.current_wf.responder.handle_exception(task_id, fl_ctx)
+                        self.current_wf.controller.communicator.handle_exception(task_id, fl_ctx)
                 return self._task_try_again()
 
             self.log_debug(fl_ctx, "firing event EventType.AFTER_TASK_DATA_FILTER")
@@ -330,7 +335,9 @@ class ServerRunner(TBI):
                     self.log_debug(fl_ctx, "no current workflow - asked client to try again later")
                     return "", "", None
 
-                task_name, task_id, task_data = self.current_wf.responder.process_task_request(client, fl_ctx)
+                task_name, task_id, task_data = self.current_wf.controller.communicator.process_task_request(
+                    client, fl_ctx
+                )
 
                 if task_name and task_name != SpecialTaskName.TRY_AGAIN:
                     if task_data:
@@ -371,7 +378,7 @@ class ServerRunner(TBI):
                 if self.current_wf is None:
                     return
 
-                self.current_wf.responder.handle_dead_job(client_name=client_name, fl_ctx=fl_ctx)
+                self.current_wf.controller.communicator.handle_dead_job(client_name=client_name, fl_ctx=fl_ctx)
             except Exception as e:
                 self.log_exception(
                     fl_ctx, f"Error processing dead job by workflow {self.current_wf.id}: {secure_format_exception(e)}"
@@ -475,7 +482,7 @@ class ServerRunner(TBI):
                 self.log_debug(fl_ctx, "firing event EventType.BEFORE_PROCESS_SUBMISSION")
                 self.fire_event(EventType.BEFORE_PROCESS_SUBMISSION, fl_ctx)
 
-                self.current_wf.responder.process_submission(
+                self.current_wf.controller.communicator.process_submission(
                     client=client, task_name=task_name, task_id=task_id, result=result, fl_ctx=fl_ctx
                 )
                 self.log_info(fl_ctx, "finished processing client result by {}".format(self.current_wf.id))
@@ -501,11 +508,11 @@ class ServerRunner(TBI):
         self.log_debug(fl_ctx, f"received task_check on task {task_id}")
 
         with self.wf_lock:
-            if self.current_wf is None or self.current_wf.responder is None:
+            if self.current_wf is None or self.current_wf.controller is None:
                 self.log_info(fl_ctx, "no current workflow - dropped task_check.")
                 return make_reply(ReturnCode.TASK_UNKNOWN)
 
-            task = self.current_wf.responder.process_task_check(task_id=task_id, fl_ctx=fl_ctx)
+            task = self.current_wf.controller.communicator.process_task_check(task_id=task_id, fl_ctx=fl_ctx)
             if task:
                 self.log_debug(fl_ctx, f"task {task_id} is still good")
                 return make_reply(ReturnCode.OK)
